Add domain and record models
Add django models for domain and record to be able to manage the
data stored in the PowerDNS database.

To avoid too many queries to the Serveradmin and PowerDNS databases
when applying changes we add two columns to the standard PowerDNS
schema (record_id, object_id) and add them as attributes to the models.

- Add domain and record models to manage PowerDNS using django ORM
- Add two attributes for fewer SQL queries at signal processing
- Add migration to add the two columns to existing PowerDNS databases